feat(psp-accounts): Address code review comments
