+++ app/services/customers/payment_provider_finder.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+module Customers
+  module PaymentProviderFinder
+    extend ActiveSupport::Concern
+
+    included do
+      def payment_provider(customer)
+        payment_provider_result = PaymentProviders::FindService.new(
+          organization_id: customer.organization_id,
+          code: customer.payment_provider_code,
+        ).call
+
+        return nil if payment_provider_result.error&.code == 'payment_provider_not_found'
+
+        payment_provider_result.raise_if_error!
+        payment_provider_result.payment_provider
+      end
+    end
+  end
+end
