+++ app/services/payment_providers/adyen_service.rb
@@ -5,10 +5,10 @@ module PaymentProviders
     WEBHOOKS_EVENTS = %w[AUTHORISATION REFUND REFUND_FAILED].freeze
 
     def create_or_update(**args)
-      payment_provider_result = PaymentProviders::FindService.new(
+      payment_provider_result = PaymentProviders::FindService.call(
         organization_id: args[:organization].id,
         code: args[:code],
-      ).call
+      )
 
       adyen_provider = if payment_provider_result.success?
         payment_provider_result.payment_provider
@@ -49,7 +49,7 @@ module PaymentProviders
       organization = Organization.find_by(id: organization_id)
       return result.service_failure!(code: 'webhook_error', message: 'Organization not found') unless organization
 
-      payment_provider_result = PaymentProviders::FindService.new(organization_id:, code:).call
+      payment_provider_result = PaymentProviders::FindService.call(organization_id:, code:)
       return payment_provider_result unless payment_provider_result.success?
 
       validator = ::Adyen::Utils::HmacValidator.new
